refactor: add singleton trait
